# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    Comprehensive REST API for TaskFlow task management system.
    
    Features:
    - JWT-based authentication with refresh tokens
    - Complete task CRUD operations with advanced filtering
    - Full-text search across tasks, categories, and tags
    - Real-time notifications and collaboration features
    - Analytics and productivity insights
    - Bulk operations and data import/export
    
    ## Authentication
    All endpoints (except auth endpoints) require a valid JWT Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Default: 100 requests per minute per user
    - Burst: 20 requests per 10 seconds
    - Headers: X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset
    
    ## Error Handling
    All errors follow a consistent format with success/error indicators, error codes, and detailed messages.
  version: 1.0.0
  contact:
    name: TaskFlow API Support
    email: api-support@taskflow.com
    url: https://docs.taskflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://taskflow.com/terms
servers:
  - url: https://api.taskflow.com/v1
    description: Production server
  - url: https://staging-api.taskflow.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with email verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "user@example.com"
              password: "securePassword123"
              first_name: "John"
              last_name: "Doe"
              timezone: "America/New_York"
              language_code: "en"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset password using token from email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ===============================
  # User Management Endpoints
  # ===============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    UserRole:
      type: string
      enum: [user, admin, super_admin]
      description: User role enumeration
    UserStatus:
      type: string
      enum: [active, inactive, suspended, pending_verification]
      description: User account status enumeration
...
    BaseResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            message:
              type: string
              description: Success message
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - error
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              type: object
              required:
                - code
                - message
              properties:
                code:
                  type: string
                  description: Error code
                message:
                  type: string
                  description: Error message
                details:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                      message:
                        type: string
                request_id:
                  type: string
                  description: Unique request identifier
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User unique identifier
        email:
          type: string
          format: email
          description: User email address
        first_name:
          type: string
          maxLength: 100
          description: User first name
        last_name:
          type: string
          maxLength: 100
          description: User last name
        display_name:
          type: string
          description: Full display name
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: Profile picture URL
        bio:
          type: string
          nullable: true
          description: User biography
        timezone:
          type: string
          default: UTC
          description: User timezone
        date_format:
          type: string
          default: YYYY-MM-DD
          description: Preferred date format
        time_format:
          type: string
          default: 24h
          description: Preferred time format
        language_code:
          type: string
          default: en
          description: Language preference
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        email_verified:
          type: boolean
          description: Email verification status
        preferences:
          type: object
          description: User preferences
        notification_settings:
          type: object
          description: Notification preferences
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    # ===============================
    # Task Schemas
    # ===============================
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - first_name
        - last_name
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password
        first_name:
          type: string
          maxLength: 100
          description: User first name
        last_name:
          type: string
          maxLength: 100
          description: User last name
        timezone:
          type: string
          default: UTC
          description: User timezone
        language_code:
          type: string
          default: en
          description: Language preference
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          description: User password
        remember_me:
          type: boolean
          default: false
          description: Extended session flag
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh token
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: User email address
    ResetPasswordRequest:
      type: object
      required:
        - token
        - new_password
      properties:
        token:
          type: string
          description: Password reset token
        new_password:
          type: string
          minLength: 8
          description: New password
...
    RegisterResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            message:
              type: string
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                  description: JWT access token
                refresh_token:
                  type: string
                  description: Refresh token
                expires_in:
                  type: integer
                  description: Token expiration in seconds
                token_type:
                  type: string
                  enum: [Bearer]
                user:
                  $ref: '#/components/schemas/User'
    RefreshTokenResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                expires_in:
                  type: integer
                token_type:
                  type: string
                  enum: [Bearer]
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid input data"
              details:
                - field: "email"
                  message: "Invalid email format"
              request_id: "req-550e8400-e29b-41d4"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "AUTHENTICATION_ERROR"
              message: "Invalid or expired token"
              request_id: "req-550e8400-e29b-41d4"
...
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...